pyproject.toml
[project]
name = "app"
version = "1.0.0"
description = "A structured FastAPI project with versioning and Supabase integration"
authors = [
    {name = "mmdsmb", email = "mmdsmb@gmail.com"}
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg[binary,pool]>=3.1.18",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "supabase>=1.0.3",
    "pytest-asyncio>=0.26.0",
    "asyncpg>=0.30.0",
    "httpx>=0.25.0",
    "pulp>=2.8.0",
    "pandas>=2.0.0",
    "geopy>=2.0.0",
    "gdown>=4.7.1",
    "google-auth>=2.16.0",
    "google-api-python-client>=2.84.0",
    "openpyxl>=3.1.2",  # Nécessaire pour pd.read_excel()
    "email-validator>=1.3.1"
]


[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "flake8>=6.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3 